Rust 1.6

Rust Documentation

Welcome to the Rust documentation! You can use the section headings above to jump to any particular section.

1 Getting Started

If you haven't seen Rust at all yet, the first thing you should read is the introduction to The Rust Programming Language. It'll give you a good idea of what Rust is like.

The book provides a lengthy explanation of Rust, its syntax, and its concepts. Upon completing the book, you'll be an intermediate Rust developer, and will have a good grasp of the fundamental ideas behind Rust.

Rust By Example teaches you Rust through a series of small examples.

2 Language Reference

Rust does not have an exact specification yet, but an effort to describe as much of the language in as much detail as possible is in the reference.

3 Standard Library Reference

We have API documentation for the entire standard library. There's a list of crates on the left with more specific sections, or you can use the search bar at the top to search for something if you know its name.

4 The Rustonomicon

The Rustonomicon is an entire book dedicated to explaining how to write unsafe Rust code. It is for advanced Rust programmers.

5 Tools

Cargo is the Rust package manager providing access to libraries beyond the standard one, and its website contains lots of good documentation.

rustdoc is the Rust's documentation generator, a tool converting annotated source code into HTML docs.

6 FAQs

There are questions that are asked quite often, so we've made FAQs for them:

Language Design FAQ

Language FAQ

Project FAQ

How to submit a bug report

7 The Error Index

If you encounter an error while compiling your code you may be able to look it up in the Rust Compiler Error Index.

8 Community Translations

Several projects have been started to translate the documentation into other languages:

Russian

Korean

Chinese

Spanish

German